fix: ZXing.Common.BitMatrix Range Check Error - https://github.com/Spelt/ZXing.Delphi/issues/104

+++ Lib/Classes/Common/ZXing.Common.BitMatrix.pas
@@ -92,7 +92,7 @@ implementation
 
 function TBitMatrix.getBit(x, y: Integer): Boolean;
 var
-  offset, v, bits, shift: Integer;
+  offset, v, bits, shift: Int64;
   uBits: Cardinal;
 begin
   offset := y * FrowSize + TMathUtils.Asr(x, 5);
